factor out single dimension viz
