+++ src/single-dimension-viz/index.js
@@ -1,0 +1,1 @@
+export { default } from "./SingleDimensionViz";
