+++ src/viz-supervision-revocation/VizSupervisionRevocation.js
@@ -1,80 +1,9 @@
 import { ascending, group } from "d3-array";
 import PropTypes from "prop-types";
 import React, { useState, useEffect } from "react";
-import Measure from "react-measure";
-import styled from "styled-components";
-import BubbleChart from "../bubble-chart";
-import {
-  DIMENSION_KEYS,
-  THEME,
-  TOTAL_KEY,
-  VIOLATION_LABELS,
-  VIOLATION_COUNT_KEYS,
-} from "../constants";
-import ProportionalBar from "../proportional-bar";
-import { formatDemographicValue, getDimensionalBreakdown } from "../utils";
-
-const SECTION_HEIGHT = 450;
-const GUTTER = 42;
-
-const VizSupervisionRevocationWrapper = styled.div`
-  width: 100%;
-`;
-
-const BreakdownBarWrapper = styled.div`
-  height: ${(props) => props.height}px;
-  padding-bottom: 16px;
-  position: relative;
-  z-index: ${(props) => props.theme.zIndex.base + props.stackOrder};
-`;
-
-function Breakdowns({ data, dimension }) {
-  const breakdownHeight = SECTION_HEIGHT / data.size;
-  return Array.from(data, ([key, value], i) => (
-    <BreakdownBarWrapper key={key} stackOrder={data.size - i}>
-      <ProportionalBar
-        data={value}
-        height={breakdownHeight - GUTTER}
-        title={formatDemographicValue(key, dimension)}
-        showLegend={i === data.size - 1}
-      />
-    </BreakdownBarWrapper>
-  ));
-}
-
-function VizSupervisionRevocation({ currentMonthData, dimension }) {
-  return (
-    <Measure bounds>
-      {({
-        measureRef,
-        contentRect: {
-          bounds: { width },
-        },
-      }) => {
-        return (
-          <VizSupervisionRevocationWrapper ref={measureRef}>
-            {dimension === DIMENSION_KEYS.total ? (
-              width && (
-                <BubbleChart
-                  data={currentMonthData.get(TOTAL_KEY)}
-                  height={SECTION_HEIGHT}
-                  width={width}
-                />
-              )
-            ) : (
-              <Breakdowns data={currentMonthData} dimension={dimension} />
-            )}
-          </VizSupervisionRevocationWrapper>
-        );
-      }}
-    </Measure>
-  );
-}
-
-VizSupervisionRevocation.propTypes = {
-  currentMonthData: PropTypes.instanceOf(Map).isRequired,
-  dimension: PropTypes.string.isRequired,
-};
+import { THEME, VIOLATION_LABELS, VIOLATION_COUNT_KEYS } from "../constants";
+import SingleDimensionViz from "../single-dimension-viz";
+import { getDimensionalBreakdown } from "../utils";
 
 function typeCastRecords(records) {
   return records.map((record) => ({
@@ -114,7 +43,6 @@ export default function VizSupervisionRevocationContainer({
   month,
   setMonthList,
   dimension,
-  ...passThruProps
 }) {
   const [monthlyData] = useState(groupByMonth(supervisionRevocationByMonth));
 
@@ -133,13 +61,7 @@ export default function VizSupervisionRevocationContainer({
     }).map(({ id, record }) => [id, splitByViolationType(record)])
   );
 
-  return (
-    <VizSupervisionRevocation
-      currentMonthData={currentMonthData}
-      dimension={dimension}
-      {...passThruProps}
-    />
-  );
+  return <SingleDimensionViz data={currentMonthData} dimension={dimension} />;
 }
 
 VizSupervisionRevocationContainer.propTypes = {
